working backend

+++ src/main/java/project/bookcrossing/service/MessageService.java
@@ -6,10 +6,12 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import project.bookcrossing.entity.Conversation;
 import project.bookcrossing.entity.Message;
+import project.bookcrossing.entity.User;
 import project.bookcrossing.repository.MessageRepository;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class MessageService {
@@ -17,9 +19,9 @@ public class MessageService {
 	@Autowired
 	private MessageRepository messageRepository;
 
-	public ResponseEntity<Message> createMessage(Message message, Conversation conversation) {
+	public ResponseEntity<Message> createMessage(Message message, User user, Conversation conversation) {
 		try {
-			Message _message = messageRepository.save(new Message(message.getContent(), conversation));
+			Message _message = messageRepository.save(new Message(message.getContent(), user, conversation));
 			return new ResponseEntity<>(_message, HttpStatus.CREATED);
 		} catch (Exception e) {
 			return new ResponseEntity<>(null, HttpStatus.EXPECTATION_FAILED);
@@ -27,7 +29,7 @@ public class MessageService {
 	}
 
 	public ResponseEntity<List<Message>> getMessagesByConversation(Conversation conversation) {
-		List<Message> messages = new ArrayList<>(messageRepository.findByConversationOrderByDateAsc(conversation));
+		List<Message> messages = new ArrayList<>(messageRepository.findByConversationOrderByDateDesc(conversation));
 		if (!messages.isEmpty()) {
 			return new ResponseEntity<>(messages, HttpStatus.OK);
 		} else {
@@ -46,18 +48,34 @@ public class MessageService {
 
 	public ResponseEntity<HttpStatus> deleteMessage(long message_id){
 		try {
-			messageRepository.deleteById(message_id);
-			return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+			if (this.setUser(message_id).getStatusCode().equals(HttpStatus.OK)) {
+				messageRepository.deleteById(message_id);
+				return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+			}
+			return new ResponseEntity<>(HttpStatus.EXPECTATION_FAILED);
 		} catch (Exception e) {
 			return new ResponseEntity<>(HttpStatus.EXPECTATION_FAILED);
 		}
 	}
 
+	private ResponseEntity<Message> setUser(long messageId) {
+		Optional<Message> _message = messageRepository.findByIdMessage(messageId);
+		if (_message.isPresent()) {
+			Message message = _message.get();
+			message.setSender(null);
+			return new ResponseEntity<>(messageRepository.save(message), HttpStatus.OK);
+		} else {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		}
+	}
+
 	public ResponseEntity<HttpStatus> deleteByConversation(Conversation conversation){
 		List<Message> messages = messageRepository.getAllByConversation(conversation);
-		if (messages != null && !messages.isEmpty()) {
-			for (Message item : messages) {
-				deleteMessage(item.getId_message());
+		if (messages != null) {
+			if (!messages.isEmpty()) {
+				for (Message item : messages) {
+					deleteMessage(item.getId_message());
+				}
 			}
 			return new ResponseEntity<>(HttpStatus.NO_CONTENT);
 		}
